+++ src/com/shtm/po/RegionExample.java
@@ -1,5 +1,6 @@
 package com.shtm.po;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -114,52 +115,52 @@ public class RegionExample {
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdEqualTo(Double value) {
+        public Criteria andRegionIdEqualTo(BigDecimal value) {
             addCriterion("REGION_ID =", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdNotEqualTo(Double value) {
+        public Criteria andRegionIdNotEqualTo(BigDecimal value) {
             addCriterion("REGION_ID <>", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdGreaterThan(Double value) {
+        public Criteria andRegionIdGreaterThan(BigDecimal value) {
             addCriterion("REGION_ID >", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdGreaterThanOrEqualTo(Double value) {
+        public Criteria andRegionIdGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_ID >=", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdLessThan(Double value) {
+        public Criteria andRegionIdLessThan(BigDecimal value) {
             addCriterion("REGION_ID <", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdLessThanOrEqualTo(Double value) {
+        public Criteria andRegionIdLessThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_ID <=", value, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdIn(List<Double> values) {
+        public Criteria andRegionIdIn(List<BigDecimal> values) {
             addCriterion("REGION_ID in", values, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdNotIn(List<Double> values) {
+        public Criteria andRegionIdNotIn(List<BigDecimal> values) {
             addCriterion("REGION_ID not in", values, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdBetween(Double value1, Double value2) {
+        public Criteria andRegionIdBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_ID between", value1, value2, "regionId");
             return (Criteria) this;
         }
 
-        public Criteria andRegionIdNotBetween(Double value1, Double value2) {
+        public Criteria andRegionIdNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_ID not between", value1, value2, "regionId");
             return (Criteria) this;
         }
@@ -314,52 +315,52 @@ public class RegionExample {
             return (Criteria) this;
         }
 
-        public Criteria andParentIdEqualTo(Double value) {
+        public Criteria andParentIdEqualTo(BigDecimal value) {
             addCriterion("PARENT_ID =", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdNotEqualTo(Double value) {
+        public Criteria andParentIdNotEqualTo(BigDecimal value) {
             addCriterion("PARENT_ID <>", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdGreaterThan(Double value) {
+        public Criteria andParentIdGreaterThan(BigDecimal value) {
             addCriterion("PARENT_ID >", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdGreaterThanOrEqualTo(Double value) {
+        public Criteria andParentIdGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("PARENT_ID >=", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdLessThan(Double value) {
+        public Criteria andParentIdLessThan(BigDecimal value) {
             addCriterion("PARENT_ID <", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdLessThanOrEqualTo(Double value) {
+        public Criteria andParentIdLessThanOrEqualTo(BigDecimal value) {
             addCriterion("PARENT_ID <=", value, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdIn(List<Double> values) {
+        public Criteria andParentIdIn(List<BigDecimal> values) {
             addCriterion("PARENT_ID in", values, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdNotIn(List<Double> values) {
+        public Criteria andParentIdNotIn(List<BigDecimal> values) {
             addCriterion("PARENT_ID not in", values, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdBetween(Double value1, Double value2) {
+        public Criteria andParentIdBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("PARENT_ID between", value1, value2, "parentId");
             return (Criteria) this;
         }
 
-        public Criteria andParentIdNotBetween(Double value1, Double value2) {
+        public Criteria andParentIdNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("PARENT_ID not between", value1, value2, "parentId");
             return (Criteria) this;
         }
@@ -374,52 +375,52 @@ public class RegionExample {
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelEqualTo(Double value) {
+        public Criteria andRegionLevelEqualTo(BigDecimal value) {
             addCriterion("REGION_LEVEL =", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelNotEqualTo(Double value) {
+        public Criteria andRegionLevelNotEqualTo(BigDecimal value) {
             addCriterion("REGION_LEVEL <>", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelGreaterThan(Double value) {
+        public Criteria andRegionLevelGreaterThan(BigDecimal value) {
             addCriterion("REGION_LEVEL >", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelGreaterThanOrEqualTo(Double value) {
+        public Criteria andRegionLevelGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_LEVEL >=", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelLessThan(Double value) {
+        public Criteria andRegionLevelLessThan(BigDecimal value) {
             addCriterion("REGION_LEVEL <", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelLessThanOrEqualTo(Double value) {
+        public Criteria andRegionLevelLessThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_LEVEL <=", value, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelIn(List<Double> values) {
+        public Criteria andRegionLevelIn(List<BigDecimal> values) {
             addCriterion("REGION_LEVEL in", values, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelNotIn(List<Double> values) {
+        public Criteria andRegionLevelNotIn(List<BigDecimal> values) {
             addCriterion("REGION_LEVEL not in", values, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelBetween(Double value1, Double value2) {
+        public Criteria andRegionLevelBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_LEVEL between", value1, value2, "regionLevel");
             return (Criteria) this;
         }
 
-        public Criteria andRegionLevelNotBetween(Double value1, Double value2) {
+        public Criteria andRegionLevelNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_LEVEL not between", value1, value2, "regionLevel");
             return (Criteria) this;
         }
@@ -434,52 +435,52 @@ public class RegionExample {
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderEqualTo(Double value) {
+        public Criteria andRegionOrderEqualTo(BigDecimal value) {
             addCriterion("REGION_ORDER =", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderNotEqualTo(Double value) {
+        public Criteria andRegionOrderNotEqualTo(BigDecimal value) {
             addCriterion("REGION_ORDER <>", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderGreaterThan(Double value) {
+        public Criteria andRegionOrderGreaterThan(BigDecimal value) {
             addCriterion("REGION_ORDER >", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderGreaterThanOrEqualTo(Double value) {
+        public Criteria andRegionOrderGreaterThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_ORDER >=", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderLessThan(Double value) {
+        public Criteria andRegionOrderLessThan(BigDecimal value) {
             addCriterion("REGION_ORDER <", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderLessThanOrEqualTo(Double value) {
+        public Criteria andRegionOrderLessThanOrEqualTo(BigDecimal value) {
             addCriterion("REGION_ORDER <=", value, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderIn(List<Double> values) {
+        public Criteria andRegionOrderIn(List<BigDecimal> values) {
             addCriterion("REGION_ORDER in", values, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderNotIn(List<Double> values) {
+        public Criteria andRegionOrderNotIn(List<BigDecimal> values) {
             addCriterion("REGION_ORDER not in", values, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderBetween(Double value1, Double value2) {
+        public Criteria andRegionOrderBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_ORDER between", value1, value2, "regionOrder");
             return (Criteria) this;
         }
 
-        public Criteria andRegionOrderNotBetween(Double value1, Double value2) {
+        public Criteria andRegionOrderNotBetween(BigDecimal value1, BigDecimal value2) {
             addCriterion("REGION_ORDER not between", value1, value2, "regionOrder");
             return (Criteria) this;
         }
